Changed the way stats work : floats around 1 rather than arbitrary ints. Attack power's calculation modified accordingly. Enhanced constructors, fixed fear status issues, enabled the possibility to check the animals' stats while choosing an attack

+++ src/Action/Attack/Abstract/IAttack.java
@@ -5,7 +5,7 @@ import Animal.Creation.Abstract.IAnimal;
 public interface IAttack {
     IAnimal getAttackOwner();
     void setAttackOwner(IAnimal attackOwner);
-    void performAttack(IAnimal target, int damage);
+    void performAttack(IAnimal target, float damage);
     String getAttackName();
     int getDamageBase();
     float getAccuracy();
